Element: Move api::element to its own module
This commt moves the Element, ElementLink and it's related tests into
a new file, element.rs, that lives under the api directory.  This is better
for clarity in the crate, and help prevent merge conflicts.
